Add portfolio feature, improve sync performance, and enhance UI
- Add portfolio tracking system (models, routes, database migrations)
- Implement lazy sync strategy (1-hour threshold vs continuous polling)
- Enhance public API with portfolio endpoints
- Expand frontend with portfolio views and improved styling
- Add market cap population and hosting price sync utilities

+++ backend/app/models/company.py
@@ -3,7 +3,7 @@
 from datetime import datetime
 from uuid import uuid4
 
-from sqlalchemy import Boolean, Integer, String, DateTime, Index
+from sqlalchemy import BigInteger, Boolean, Integer, String, DateTime, Index
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -25,7 +25,7 @@ class Company(Base):
     name: Mapped[str] = mapped_column(String(255), nullable=False)
     sector: Mapped[str | None] = mapped_column(String(100))
     industry_group: Mapped[str | None] = mapped_column(String(100))
-    market_cap: Mapped[int | None] = mapped_column(Integer)
+    market_cap: Mapped[int | None] = mapped_column(BigInteger)
     is_asx200: Mapped[bool] = mapped_column(Boolean, default=False)
     is_asx300: Mapped[bool] = mapped_column(Boolean, default=False)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
